AuthenticationService is fully completed

+++ src/main/scala/com/itechart/project/routes/AttachmentRoutes.scala
@@ -2,7 +2,10 @@ package com.itechart.project.routes
 
 import cats.effect.Sync
 import cats.implicits._
-import com.itechart.project.services.AttachmentService
+import com.itechart.project.domain.user.Role
+import com.itechart.project.dto.auth.LoggedInUser
+import com.itechart.project.routes.access.AccessChecker.isResourceAvailable
+import com.itechart.project.services.{AttachmentService, CategoryService}
 import com.itechart.project.services.error.AttachmentErrors.AttachmentFileError
 import com.itechart.project.services.error.AttachmentErrors.AttachmentFileError.{
   AttachmentNotFound,
@@ -11,7 +14,8 @@ import com.itechart.project.services.error.AttachmentErrors.AttachmentFileError.
 import fs2.io.file.Files
 import org.http4s.dsl.Http4sDsl
 import org.http4s.circe.CirceEntityCodec.circeEntityEncoder
-import org.http4s.{EntityEncoder, HttpRoutes, Response, StaticFile}
+import org.http4s.circe.JsonDecoder
+import org.http4s.{AuthedRoutes, EntityEncoder, HttpRoutes, Response, StaticFile}
 import org.typelevel.log4cats.Logger
 
 import scala.util.Try
@@ -34,42 +38,62 @@ object AttachmentRoutes {
       } yield response
     }
 
-    def deleteCategory(): HttpRoutes[F] = HttpRoutes.of[F] { case DELETE -> Root / "attachments" / LongVar(id) =>
-      val res = for {
-        deleted <- attachmentService.deleteFile(id)
-      } yield deleted
-
-      marshalResponse(res)
-    }
-
     object LongVar {
       def unapply(value: String): Option[Long] = Try(value.toLong).toOption
     }
 
-    def attachmentErrorToHttpResponse(error: AttachmentFileError): F[Response[F]] = {
-      error match {
-        case e: AttachmentNotFound    => NotFound(e.message)
-        case e: InvalidItemAttachment => Conflict(e.message)
+    getAttachment
+  }
 
-        case e => BadRequest(e.message)
-      }
+  def securedRoutes[F[_]: Sync: Logger: JsonDecoder](
+    attachmentService: AttachmentService[F]
+  ): AuthedRoutes[LoggedInUser, F] = {
+    val dsl = new Http4sDsl[F] {}
+    import dsl._
+
+    def deleteCategory(): AuthedRoutes[LoggedInUser, F] = AuthedRoutes.of {
+      case DELETE -> Root / "attachments" / LongVar(id) as user =>
+        if (!isResourceAvailable(user.value.role, List(Role.Manager))) Forbidden()
+        else {
+          val res = for {
+            deleted <- attachmentService.deleteFile(id)
+          } yield deleted
+
+          marshalResponse(res)
+        }
     }
 
-    def marshalResponse[T](
-      result: F[Either[AttachmentFileError, T]]
-    )(
-      implicit E: EntityEncoder[F, T]
-    ): F[Response[F]] =
-      result
-        .flatMap {
-          case Left(error) => attachmentErrorToHttpResponse(error) <* Logger[F].info("ERROR: " + error.message)
-          case Right(dto)  => Ok(dto)
-        }
-        .handleErrorWith { ex =>
-          InternalServerError(ex.getMessage) <* Logger[F].error(ex.getMessage)
-        }
+    deleteCategory()
+  }
 
-    getAttachment <+> deleteCategory()
+  def attachmentErrorToHttpResponse[F[_]: Sync: Logger](error: AttachmentFileError): F[Response[F]] = {
+    val dsl = new Http4sDsl[F] {}
+    import dsl._
+
+    error match {
+      case e: AttachmentNotFound    => NotFound(e.message)
+      case e: InvalidItemAttachment => Conflict(e.message)
+
+      case e => BadRequest(e.message)
+    }
+  }
+
+  def marshalResponse[F[_]: Sync: Logger, T](
+    result: F[Either[AttachmentFileError, T]]
+  )(
+    implicit E: EntityEncoder[F, T]
+  ): F[Response[F]] = {
+    val dsl = new Http4sDsl[F] {}
+    import dsl._
+
+    result
+      .flatMap {
+        case Left(error) => attachmentErrorToHttpResponse(error) <* Logger[F].info("ERROR: " + error.message)
+        case Right(dto)  => Ok(dto)
+      }
+      .handleErrorWith { ex =>
+        InternalServerError(ex.getMessage) <* Logger[F].error(ex.getMessage)
+      }
   }
 
 }
